requirements.txt
numpy==1.26.4
pillow==10.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
six==1.16.0
wxPython==4.2.1
